Moved back to es6 target with commonjs module resolution and graphql fix
The move to es5 is not needed when the module resolution is set to
commonjs because then "require" is used instead of "import".

Then we need to move the graphql files as they are not moved anymore.
That is because they would be moved with babel as we have moved to
tsc for transpiling for typscript support. This is fixed with the package
cpx with which the files get all moved to the right folder.